Move logic inside the script sync_tables script
